+++ mnelab/dialogs/__init__.py
@@ -18,13 +18,16 @@ from .interpolatebadsdialog import InterpolateBadsDialog
 from .metainfodialog import MetaInfoDialog
 from .montagedialog import MontageDialog
 from .pickchannelsdialog import PickChannelsDialog
+from .referencebipolardialog import ReferenceBipolarDialog
 from .referencedialog import ReferenceDialog
 from .runicadialog import RunICADialog
 from .xdfstreamsdialog import XDFStreamsDialog
 
 
-__all__ = [AnnotationsDialog, AppendDialog, CalcDialog,
-           ChannelPropertiesDialog, CropDialog, EpochDialog, ERDSDialog,
-           ErrorMessageBox, EventsDialog, FilterDialog, FindEventsDialog,
-           HistoryDialog, InterpolateBadsDialog, MetaInfoDialog, MontageDialog,
-           PickChannelsDialog, ReferenceDialog, RunICADialog, XDFStreamsDialog]
+__all__ = [AnnotationsDialog, CalcDialog, AppendDialog, ERDSDialog,
+           ChannelPropertiesDialog, CropDialog,
+           EpochDialog, ErrorMessageBox, EventsDialog, FilterDialog,
+           FindEventsDialog, HistoryDialog, InterpolateBadsDialog,
+           MetaInfoDialog, MontageDialog, PickChannelsDialog,
+           ReferenceBipolarDialog, ReferenceDialog, RunICADialog,
+           XDFStreamsDialog]