+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 ### Changed
 - PySide2 installed by default if wrapped Qt bindings (PyQt5, PySide2) are not found ([#187](https://github.com/cbrnr/mnelab/pull/187) by [Guillaume Dollé](https://github.com/gdolle))
+- Add support for bipolar reference [#173](https://github.com/cbrnr/mnelab/pull/173) by [Guillaume Dollé](https://github.com/gdolle))
 
 ### Added
 - Add Python 3.9 support ([#190](https://github.com/cbrnr/mnelab/pull/190) by [Clemens Brunner](https://github.com/cbrnr))